Crate wasmtime_internal_cache

⚠️ Warning ⚠️: this crate is an internal-only crate for the Wasmtime project and is not intended for general use. APIs are not strictly reviewed for safety and usage outside of Wasmtime may have bugs. If you’re interested in using this feel free to file an issue on the Wasmtime repository to start a discussion about doing so, but otherwise be aware that your usage of this crate is not supported.

Structs

Cache

Global configuration for how the cache is managed

CacheConfig

Global configuration for how the cache is managed

ModuleCacheEntry

Module level cache entry.

Functions

create_new_config

Creates a new configuration file at specified path, or default path if None is passed. Fails if file already exists.