wasmtime_internal_cache

Struct CacheConfig

pub struct CacheConfig { /* private fields */ }

Global configuration for how the cache is managed

Implementations

impl CacheConfig

pub fn new() -> Self

Creates a new set of configuration which represents a disabled cache

pub fn from_file(config_file: Option<&Path>) -> Result<Self>

Loads cache configuration specified at path.

This method will read the file specified by path on the filesystem and attempt to load cache configuration from it. This method can also fail due to I/O errors, misconfiguration, syntax errors, etc. For expected syntax in the configuration file see the documentation online.

Passing in None loads cache configuration from the system default path. This is located, for example, on Unix at $HOME/.config/wasmtime/config.toml and is typically created with the wasmtime config new command.

Errors

This method can fail due to any error that happens when loading the file pointed to by path and attempting to load the cache configuration.

pub fn worker_event_queue_size(&self) -> u64

Returns worker_event_queue_size.

Panics if the cache is disabled.

pub fn baseline_compression_level(&self) -> i32

Returns baseline_compression_level.

Panics if the cache is disabled.

pub fn optimized_compression_level(&self) -> i32

Returns optimized_compression_level.

Panics if the cache is disabled.

pub fn optimized_compression_usage_counter_threshold(&self) -> u64

Returns optimized_compression_usage_counter_threshold.

Panics if the cache is disabled.

pub fn cleanup_interval(&self) -> Duration

Returns cleanup_interval.

Panics if the cache is disabled.

pub fn optimizing_compression_task_timeout(&self) -> Duration

Returns optimizing_compression_task_timeout.

Panics if the cache is disabled.

pub fn allowed_clock_drift_for_files_from_future(&self) -> Duration

Returns allowed_clock_drift_for_files_from_future.

Panics if the cache is disabled.

pub fn file_count_soft_limit(&self) -> u64

Returns file_count_soft_limit.

Panics if the cache is disabled.

pub fn files_total_size_soft_limit(&self) -> u64

Returns files_total_size_soft_limit.

Panics if the cache is disabled.

pub fn file_count_limit_percent_if_deleting(&self) -> u8

Returns file_count_limit_percent_if_deleting.

Panics if the cache is disabled.

pub fn files_total_size_limit_percent_if_deleting(&self) -> u8

Returns files_total_size_limit_percent_if_deleting.

Panics if the cache is disabled.

pub fn directory(&self) -> &PathBuf

Returns path to the cache directory.

Panics if the cache is disabled.

pub fn with_directory(&mut self, directory: impl Into<PathBuf>) -> &mut Self

Specify where the cache directory is. Must be an absolute path.

pub fn with_worker_event_queue_size(&mut self, size: u64) -> &mut Self

Size of cache worker event queue. If the queue is full, incoming cache usage events will be dropped.

pub fn with_baseline_compression_level(&mut self, level: i32) -> &mut Self

Compression level used when a new cache file is being written by the cache system. Wasmtime uses zstd compression.

pub fn with_optimized_compression_level(&mut self, level: i32) -> &mut Self

Compression level used when the cache worker decides to recompress a cache file. Wasmtime uses zstd compression.

pub fn with_optimized_compression_usage_counter_threshold( &mut self, threshold: u64, ) -> &mut Self

One of the conditions for the cache worker to recompress a cache file is to have usage count of the file exceeding this threshold.

pub fn with_cleanup_interval(&mut self, interval: Duration) -> &mut Self

When the cache worker is notified about a cache file being updated by the cache system and this interval has already passed since last cleaning up, the worker will attempt a new cleanup.

pub fn with_optimizing_compression_task_timeout( &mut self, timeout: Duration, ) -> &mut Self

When the cache worker decides to recompress a cache file, it makes sure that no other worker has started the task for this file within the last optimizing-compression-task-timeout interval. If some worker has started working on it, other workers are skipping this task.

pub fn with_allowed_clock_drift_for_files_from_future( &mut self, drift: Duration, ) -> &mut Self

Locks

When the cache worker attempts acquiring a lock for some task, it checks if some other worker has already acquired such a lock. To be fault tolerant and eventually execute every task, the locks expire after some interval. However, because of clock drifts and different timezones, it would happen that some lock was created in the future. This setting defines a tolerance limit for these locks. If the time has been changed in the system (i.e. two years backwards), the cache system should still work properly. Thus, these locks will be treated as expired (assuming the tolerance is not too big).

Cache files

Similarly to the locks, the cache files or their metadata might have modification time in distant future. The cache system tries to keep these files as long as possible. If the limits are not reached, the cache files will not be deleted. Otherwise, they will be treated as the oldest files, so they might survive. If the user actually uses the cache file, the modification time will be updated.

pub fn with_file_count_soft_limit(&mut self, limit: u64) -> &mut Self

Soft limit for the file count in the cache directory.

This doesnโt include files with metadata. To learn more, please refer to the cache system section.

pub fn with_files_total_size_soft_limit(&mut self, limit: u64) -> &mut Self

Soft limit for the total size* of files in the cache directory.

This doesnโt include files with metadata. To learn more, please refer to the cache system section.

*this is the file size, not the space physically occupied on the disk.

pub fn with_file_count_limit_percent_if_deleting( &mut self, percent: u8, ) -> &mut Self

If file-count-soft-limit is exceeded and the cache worker performs the cleanup task, then the worker will delete some cache files, so after the task, the file count should not exceed file-count-soft-limit * file-count-limit-percent-if-deleting.

This doesnโt include files with metadata. To learn more, please refer to the cache system section.

pub fn with_files_total_size_limit_percent_if_deleting( &mut self, percent: u8, ) -> &mut Self

If files-total-size-soft-limit is exceeded and cache worker performs the cleanup task, then the worker will delete some cache files, so after the task, the files total size should not exceed files-total-size-soft-limit * files-total-size-limit-percent-if-deleting.

This doesnโt include files with metadata. To learn more, please refer to the cache system section.